% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BustoolsMetadata.R
\docType{class}
\name{BustoolsMetadata-class}
\alias{BustoolsMetadata-class}
\alias{BustoolsMetadata}
\title{BustoolsMetadata S4 class}
\description{
An S4 class that allows to get the required metadata to download, install and 
run bustools to generate a count matrix for droplet-based single-cell RNA-seq data.
}
\section{Slots}{

\describe{
\item{\code{download_bustools}}{A logical allowing to use an already installed version}

\item{\code{bustools_windows_url}}{URL to the binary of bustools for windows}

\item{\code{bustools_linux_url}}{URL to the binary of bustools for linux}

\item{\code{bustools_osx_url}}{URL to the binary of bustools for MacOS}

\item{\code{bustools_dir}}{Name of the directory where bustools will be installed}

\item{\code{unix_bustools_name}}{Name of the bustools executable in linux and macOS}

\item{\code{windows_bustools_name}}{Name of the bustools executable in windows}
}}

