% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcriptToGene.R
\name{create_tx2gene}
\alias{create_tx2gene}
\title{Create transcript to gene mapping file}
\usage{
create_tx2gene(myAbundanceMetadata, myBgeeMetadata, myUserMetadata)
}
\arguments{
\item{myAbundanceMetadata}{A descendant object of the Class myAbundanceMetadata.}

\item{myBgeeMetadata}{A Reference Class BgeeMetadata object.}

\item{myUserMetadata}{A Reference Class UserMetadata object.}
}
\description{
Create transcript to gene mapping file as used by tximport. 
The file contains both genic and intergenic regions.
}
\author{
Julien Wollbrett
}
