% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kallisto.R
\name{download_kallisto}
\alias{download_kallisto}
\title{Download binary version of kallisto.}
\usage{
download_kallisto(myKallistoMetadata, myUserMetadata)
}
\arguments{
\item{myKallistoMetadata}{A Reference Class KallistoMetadata object.}

\item{myUserMetadata}{A Reference Class UserMetadata object.}
}
\description{
Check your OS and download correct binary 
version of kallisto.
}
\examples{
{
  kallisto <- new("KallistoMetadata")
  user <- new("UserMetadata")
  download_kallisto(kallisto, user)
}

}
\author{
Julien Wollbrett.
}
