% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateIntergenicRegions.R
\name{generate_initial_intergenic_regions}
\alias{generate_initial_intergenic_regions}
\title{generates the initial intergenic regions}
\usage{
generate_initial_intergenic_regions(
  gene_gtf_path = "./genomes/Homo_sapiens.GRCh38.gtf.gz",
  genome_fasta_path = "./genomes/Homo_sapiens.GRCh38.genome.fa",
  N_block_size = 31,
  N_proportion = 0.05,
  output_gtf_path = "./intergenic_regions/",
  Rout_path = "./"
)
}
\arguments{
\item{gene_gtf_path}{full path to input gene gtf file}

\item{genome_fasta_path}{full path to input genome fasta file}

\item{N_block_size}{number of successive N from which it is considered as a block of N (and removed)}

\item{N_proportion}{higher proportion of N than this threshold results in removing the sequence}

\item{output_gtf_path}{full path to output folder + base name for output files}

\item{Rout_path}{path to  output .Rout file (optional)}

\item{download_gtf_fasta}{boolean on whether to download the gtf and fasta from ensembl or}
}
\description{
generates intergenic regions from the species GTF and FASTA file by taking intergenic regions upstream and downstream of each gene
}
\author{
Alessandro Brandulas Cammarata

Julien Wollbrett

Julien Roux

Marta Rosikiewicz
}
