% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateIntergenicRegions.R
\name{generate_reference_intergenic_regions}
\alias{generate_reference_intergenic_regions}
\title{Generates reference intergenic regions by removing outlier intergenics}
\usage{
generate_reference_intergenic_regions(
  intergenic_regions_path,
  tx2gene_path,
  gene2biotype_path,
  TPM_path,
  output_path = "./reference_intergenic_regions",
  species = "my_species"
)
}
\arguments{
\item{intergenic_regions_path}{path to the intergenic regions}

\item{tx2gene_path}{path to the tx2gene file of the species}

\item{gene2biotype_path}{path to the gene2biotype file of the species}

\item{TPM_path}{path to the abundance files of the libraries}

\item{output_path}{path to the output folder}
}
\description{
We sum the TPM values of intergenic regions in specified directory containing multiple libraries and find a threshold to remove outlier intergenic regions based on protein coding genes
we then filter the fasta file containing all intergenic regions to the reference intergenic regions by removing outliers
}
\author{
Alessandro Brandulas Cammarata
}
