% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kallisto_singlecell.R
\name{get_bustools_path}
\alias{get_bustools_path}
\title{Retrieve execution path or download Bustools}
\usage{
get_bustools_path(bustools_metadata, myUserMetadata)
}
\arguments{
\item{bustools_metadata}{An object of class BustoolsMetadata containing the required metadata to download
and run bustools.}

\item{myUserMetadata}{An object of class UserMetadata containing user-specific metadata such as working directory.}
}
\value{
The absolute path to the bustools executable.
}
\description{
Checks if bustools binary already exists in the system or if absent and download flag
is set to TRUE, it downloads bustools from github.
}
