% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_intergenic_release_path}
\alias{get_intergenic_release_path}
\title{Path to bgee release directory}
\usage{
get_intergenic_release_path(myBgeeMetadata, myUserMetadata)
}
\value{
path to the file containing details about intergenic releases
}
\description{
helper function to get the path to the bgee release directory
}
