% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_kallisto_dir_path}
\alias{get_kallisto_dir_path}
\title{Retrieve path to kallisto directory}
\usage{
get_kallisto_dir_path(myAbundanceMetadata, myUserMetadata)
}
\arguments{
\item{myAbundanceMetadata}{Reference Class AbundanceMetadata object.}

\item{myUserMetadata}{Reference Class UserMetadata object.}
}
\value{
path to kallisto dir
}
\description{
Retireve path to kallisto directory. This path depends
on the OS
}
