% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_output_dir}
\alias{get_output_dir}
\title{Retrieve name of output directory}
\usage{
get_output_dir(myUserMetadata)
}
\value{
Name of the output directory
}
\description{
Retireve name of output directory depending on myUserMetadata@rnaseq_lib_path and myUserMetadata@run_ids
if myUserMetadata@run_ids is empty then the output directory will correspond to name of the last directory of myUserMetadata@rnaseq_lib_path.
Otherwise the name of the output directory will be the concatenation of the name of the last directory of myUserMetadata@rnaseq_lib_path and all myUserMetadata@run_ids
}
