% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presenceAbsence.R
\name{get_ref_intergenic_ids}
\alias{get_ref_intergenic_ids}
\title{Retrieve dataframe of reference intergenic IDs}
\usage{
get_ref_intergenic_ids(myBgeeMetadata, myUserMetadata)
}
\arguments{
\item{myBgeeMetadata}{A Class BgeeMetadata object.}

\item{myUserMetadata}{A Class UserMetadata object.}
}
\value{
A dataframe containing reference intergenic ids
}
\description{
Retrieve a dataframe of reference intergenic IDs. It is generated using the reference intergenic fasta file from Bgee FTP.
}
\author{
Julien Wollbrett

Julien Roux
}
