% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSummary.R
\name{get_summary_stats}
\alias{get_summary_stats}
\title{Gather statistical information}
\usage{
get_summary_stats(userFile, outDir)
}
\arguments{
\item{userFile}{A data frame containing all information of each library}

\item{outDir}{Output directory where the generated file should be saved}
}
\value{
A tsv file
}
\description{
Collect the statistics provided by the gene_cutoff_info_file from each individual library,
in order to generate a global summary file.
}
\author{
Sara Fonseca Costa
}
