% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BgeeMetadata.R
\docType{methods}
\name{getIntergenicPrefix}
\alias{getIntergenicPrefix}
\alias{getIntergenicPrefix,BgeeMetadata-method}
\alias{getIntergenicPrefix,bgeeMetadata}
\title{`intergenic_prefix` Getter}
\usage{
getIntergenicPrefix(bgeeObject)

\S4method{getIntergenicPrefix}{BgeeMetadata}(bgeeObject)
}
\arguments{
\item{bgeeObject}{The BgeeMetadata object}
}
\value{
the value of the `intergenic_prefix` slot of the object
}
\description{
Get value of the `intergenic_prefix` slot
}
\examples{
{
bgee <- new("BgeeMetadata")
intergenic_prefix <- getIntergenicPrefix(bgee)
}

}
