% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BgeeMetadata.R
\docType{methods}
\name{getIntergenicRelease}
\alias{getIntergenicRelease}
\alias{getIntergenicRelease,BgeeMetadata-method}
\alias{getIntergenicRelease,bgeeMetadata}
\title{`intergenic_release` Getter}
\usage{
getIntergenicRelease(bgeeObject)

\S4method{getIntergenicRelease}{BgeeMetadata}(bgeeObject)
}
\arguments{
\item{bgeeObject}{The BgeeMetadata object}
}
\value{
the value of the `intergenic_release` slot of the object
}
\description{
Get value of the `intergenic_release` slot
}
\examples{
{
bgee <- new("BgeeMetadata")
intergenic_release <- getIntergenicRelease(bgee)
}

}
