% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KallistoMetadata.R
\name{getKallistoPath}
\alias{getKallistoPath}
\title{Path to kallisto tool that have to be exectuted.}
\usage{
getKallistoPath(object)
}
\arguments{
\item{myKallistoMetadata}{A Reference Class KallistoMetadata object.}
}
\value{
path to kallisto tool that have to be exectuted
}
\description{
This function retrieve the path to kallisto tool that have to be 
exectuted. 
The result will depend of the OS and of the use of a kallisto version specific 
to this pakage or not.
}
\examples{
{
myKallistoMetadata <- new("KallistoMetadata")
kallisto_exec <- getKallistoPath(myKallistoMetadata)
}

}
\author{
Julien Wollbrett.
}
