% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserMetadata.R
\docType{methods}
\name{getRunIds}
\alias{getRunIds}
\alias{getRunIds,UserMetadata-method}
\alias{getRunIds,userMetadata}
\title{`run_ids` Getter}
\usage{
getRunIds(userObject)

\S4method{getRunIds}{UserMetadata}(userObject)
}
\arguments{
\item{userObject}{The UserMetadata object}
}
\value{
the value of the `run_ids` slot of the object
}
\description{
Get value of the `run_ids` slot
}
\examples{
{
user <- new("UserMetadata")
run_ids <- getRunIds(user)
}

}
