% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserMetadata.R
\docType{methods}
\name{getSimpleArborescence}
\alias{getSimpleArborescence}
\alias{getSimpleArborescence,UserMetadata-method}
\alias{getSimpleArborescence,userMetadata}
\title{`simple_arborescence` Getter}
\usage{
getSimpleArborescence(userObject)

\S4method{getSimpleArborescence}{UserMetadata}(userObject)
}
\arguments{
\item{userObject}{The UserMetadata object}
}
\value{
the value of the `simple_arborescence` slot of the object
}
\description{
Get value of the `simple_arborescence` slot
}
\examples{
{
user <- new("UserMetadata")
simple_arborescence <- getSimpleArborescence(user)
}

}
