% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveFromCommunity.R
\name{list_community_ref_intergenic_species}
\alias{list_community_ref_intergenic_species}
\title{List species having reference intergenic sequences created by the BgeeCall community}
\usage{
list_community_ref_intergenic_species()
}
\value{
list all species having reference intergenic sequences created by the community
}
\description{
Return information related to species having reference intergenic
sequences created by the BgeeCall community
- speciesId : the NCBI species ID of the species
- url : url to the reference intergenic fasta file
- numberOfLibraries : number of libraries used to generate these reference intergenic
sequences
}
\examples{
{
list_community_ref_intergenic_species()
}


}
\author{
Julien Wollbrett
}
