% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listIntergenicRelease.R
\name{list_intergenic_release}
\alias{list_intergenic_release}
\title{List Bgee intergenic releases available to use with the BgeeCall package}
\usage{
list_intergenic_release(release = NULL)
}
\arguments{
\item{release}{A character specifying a targeted release number (e.g., "14"). If not specified, all available releases are shown.}
}
\value{
A data frame with information on Bgee intergenic releases available to use with the BgeeCall package.
}
\description{
Returns information on available Bgee intergenic releases, the access URL for FTP, and the date of release
}
\examples{
{
 list_intergenic_release()
}

}
\author{
Julien Roux, Julien Wollbrett
}
