% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listBgeeRelease.R
\name{listBgeeRelease}
\alias{listBgeeRelease}
\title{List Bgee releases available to use with bgeeCall package}
\usage{
listBgeeRelease(release = NULL)
}
\arguments{
\item{release}{A character specifying a targeted release number (e.g., "14"). If not specified, all available releases are shown.}
}
\value{
A data frame with information on Bgee releases available to use with bgeeCall package.
}
\description{
Returns information on available Bgee releases, the access URL for FTP, and the date of release
}
\examples{
{
 listBgeeRelease()
}

}
\author{
Julien Roux, Julien Wollbrett
}
