% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergingLibraries.R
\name{Pvalue_averaging}
\alias{Pvalue_averaging}
\title{Pvalue averaging}
\usage{
Pvalue_averaging(pval_collect, w_values = c(), method = "mean")
}
\arguments{
\item{pval_collect}{A data frame containing the p-values for each gene id of each library}

\item{w_values}{A vector of weights to be used in the mean p-value calculation}

\item{method}{Method used to calculate the mean p-value}
}
\value{
A dataframe containing the consensus p-value for each gene id
}
\description{
Combined p-values for each gene Id using the mean or median averaging method
}
\examples{
\dontrun{
Pvalue_averaging(pval_collect = pval_collect, w_values = c(0.5, 0.5), method = "mean")
}

}
\author{
Alessandro Brandulas Cammarata
}
