% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCallsPipeline.R
\name{run_from_dataframe}
\alias{run_from_dataframe}
\title{generate present/absent calls from a data frame}
\usage{
run_from_dataframe(myAbundanceMetadata = new("KallistoMetadata"),
  myBgeeMetadata = new("BgeeMetadata"), userMetadataDataFrame)
}
\arguments{
\item{myAbundanceMetadata}{A Reference Class BgeeMetadata object (optional)}

\item{myBgeeMetadata}{A Reference Class BgeeMetadata object (optional)}

\item{userMetadataDataFrame}{A data frame containing all information needed to 
generate UserMetadata objects. This data frame must contains 7 columns (species_id, 
run_ids, reads_size, rnaseq_lib_path, transcriptome_path, annotation_path, working_path)}
}
\value{
paths to the 4 output files generated per call generation
}
\description{
Main function allowing to generate present/absent calls for one 
library described in a data frame
}
\examples{
\dontrun{
metadata_file <- system.file("userMetadataTemplate.tsv",
                               package = "BgeeCall")
user_metadata_df <- read.table(metadata_file, header = TRUE, sep = "\\t",
                              comment.char = "#")
run_from_dataframe(userMetadataDataFrame = metadata_file)
}

}
\seealso{
run_from_file()
}
\author{
Julien Wollbrett
}
