% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCallsPipeline.R
\name{run_from_object}
\alias{run_from_object}
\title{generate present/absent calls from a UserMetadata object}
\usage{
run_from_object(myAbundanceMetadata = new("KallistoMetadata"),
  myBgeeMetadata = new("BgeeMetadata"), myUserMetadata)
}
\arguments{
\item{myAbundanceMetadata}{A Reference Class BgeeMetadata object (optional)}

\item{myBgeeMetadata}{A Reference Class BgeeMetadata object (optional)}

\item{myUserMetadata}{A Reference Class UserMetadata object.}
}
\value{
4 paths to results files
}
\description{
Main function allowing to generate present/absent calls for one 
library described as a UserMetadata object
}
\examples{
{
library(AnnotationHub)
ah <- AnnotationHub()
ah_resources <- query(ah, c("Ensembl", "Caenorhabditis elegans", "84"))
annotation_object <- ah_resources[["AH50789"]]
transcriptome_object <- rtracklayer::import.2bit(ah_resources[["AH50453"]])
user_BgeeCall <- new("UserMetadata", species_id = "6239")
# import annotation and transcriptome in the user_BgeeCall object
# it is possible to import them using an S4 object (GRanges, DNAStringSet) 
# or a file (gtf, fasta) with methods setAnnotationFromFile() and
# setTranscriptomeFromFile()
user_BgeeCall <- setAnnotationFromObject(user_BgeeCall, 
                                         annotation_object, 
                                         "WBcel235_84")
user_BgeeCall <- setTranscriptomeFromObject(user_BgeeCall, 
                                          transcriptome_object, 
                                          "WBcel235")
# provide path to the directory of your RNA-Seq library
user_BgeeCall <- setRNASeqLibPath(user_BgeeCall, 
                 system.file("extdata", "SRX099901_subset", package = "BgeeCall"))
calls_output <- run_from_object(
             myUserMetadata = user_BgeeCall)                                           
}

}
\author{
Julien Wollbrett
}
