% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BgeeMetadata.R
\docType{methods}
\name{setIntergenicRelease}
\alias{setIntergenicRelease}
\alias{setIntergenicRelease,BgeeMetadata,character-method}
\alias{setIntergenicRelease,bgeeMetadata,character}
\title{`intergenic_release` Setter}
\usage{
setIntergenicRelease(bgeeObject, intergenicRelease)

\S4method{setIntergenicRelease}{BgeeMetadata,character}(bgeeObject,
  intergenicRelease)
}
\arguments{
\item{bgeeObject}{The BgeeMetadata object}

\item{intergenicRelease}{character corresponding to the `intergenic_release`}
}
\value{
An object of the class BgeeMetadata with new `intergenic_release`
 value
}
\description{
Set value of the `intergenic_release` slot
}
\examples{
{
bgee <- new("BgeeMetadata")
bgee <- setIntergenicRelease(bgee, "0.1")
}

}
