% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserMetadata.R
\docType{methods}
\name{setOutputDir}
\alias{setOutputDir}
\alias{setOutputDir,UserMetadata,character-method}
\alias{setOutputDir,userMetadata,character}
\title{`output_dir` Setter}
\usage{
setOutputDir(userObject, outputDir)

\S4method{setOutputDir}{UserMetadata,character}(userObject, outputDir)
}
\arguments{
\item{userObject}{The UserMetadata object}

\item{outputDir}{path to the directory wanted as `output_dir`}
}
\value{
An object of the class UserMetadata with new `output_dir`
 value
}
\description{
Set value of the `output_dir` slot
}
\examples{
{
user <- new("UserMetadata")
user <- setOutputDir(user, getwd())
}

}
