% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserMetadata.R
\name{setRNASeqLibPath}
\alias{setRNASeqLibPath}
\title{`rnaseq_lib_path` Setter}
\usage{
setRNASeqLibPath(userObject, rnaSeqLibPath)
}
\arguments{
\item{userObject}{The UserMetadata object}

\item{rnaSeqLibPath}{path to the directory wanted as `rnaseq_lib_path`}
}
\value{
An object of the class UserMetadata with new `rnaseq_lib_path`
 value
}
\description{
Set value of the `rnaseq_lib_path` slot
}
\examples{
{
user <- new("UserMetadata")
user <- setRNASeqLibPath(user, getwd())
}

}
