% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserMetadata.R
\docType{methods}
\name{setSimpleArborescence}
\alias{setSimpleArborescence}
\alias{setSimpleArborescence,UserMetadata,logical-method}
\alias{setSimpleArborescence,userMetadata,logical}
\title{`simple_arborescence` Setter}
\usage{
setSimpleArborescence(userObject, simpleArborescence)

\S4method{setSimpleArborescence}{UserMetadata,logical}(userObject,
  simpleArborescence)
}
\arguments{
\item{userObject}{The UserMetadata object}

\item{simpleArborescence}{boolean defining if output files will be
created a simple arborescence (TRUE) or not (FALSE)}
}
\value{
An object of the class UserMetadata with new `simple_arborescence`
 value
}
\description{
Set value of the `simple_arborescence` slot
}
\examples{
{
user <- new("UserMetadata")
user <- setSimpleArborescence(user, FALSE)
}

}
