% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserMetadata.R
\docType{methods}
\name{setTranscriptomeFromFile,UserMetadata,character,character-method}
\alias{setTranscriptomeFromFile,UserMetadata,character,character-method}
\title{Set transcriptome_object of one UserMetadata object}
\usage{

  \S4method{setTranscriptomeFromFile}{UserMetadata,character,character}(userObject,
  transcriptomePath, transcriptomeName = "")
}
\arguments{
\item{userObject}{The UserMetadata object}

\item{annotationPath}{Absolute path to the transcriptome file}

\item{annotationName}{(optional) Name of the trancriptome Will be used to 
create folders.}
}
\value{
An object of the class UserMetadata
}
\description{
Method of the class UserMetadata. Set transcriptome_object of 
one UserMetadata object 
by providing the path to a fasta transcriptome file.
}
\details{
If no annotationNAme is provided the name of the annotation file 
will be used to create folders.
}
\examples{
{
library(AnnotationHub)
ah <- AnnotationHub()
ah_transcriptomes <- query(ah, c("FaFile","Ensembl", 
      "Caenorhabditis elegans", "Caenorhabditis_elegans.WBcel235"))
user <- new("UserMetadata")
user <- setTranscriptomeFromFile(user, ah_transcriptomes[["AH49057"]]$path,
                                 "transcriptome_name")
}

}
