% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserMetadata.R
\docType{methods}
\name{setWorkingPath}
\alias{setWorkingPath}
\title{`working_path` Setter}
\usage{
setWorkingPath(userObject, workingPath)
}
\arguments{
\item{userObject}{The UserMetadata object}

\item{workingPath}{path to the directory wanted as `working_path`}
}
\value{
An object of the class UserMetadata with new `working_path`
 value
}
\description{
Set value of the `working_path` slot
}
\examples{
{
user <- new("UserMetadata")
user <- setWorkingPath(user, getwd())
}

}
