% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BgeeMetadata.R
\docType{class}
\name{BgeeMetadata-class}
\alias{BgeeMetadata-class}
\alias{BgeeMetadata}
\title{BgeeMetadata S4 class}
\description{
An S4 class that contains all information to retrieve
intergenic regions generated by Bgee.
}
\section{Slots}{

\describe{
\item{\code{intergenic_release}}{Bgee intergenic release that will be used}

\item{\code{all_releases}}{list of all reference intergenic releases that can be
used to generate your present/absent expression calls.}

\item{\code{intergenic_prefix}}{String used to generate an intergenic release
specific output directory}
}}

