% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BgeeMetadata.R
\docType{class}
\name{BgeeMetadata-class}
\alias{BgeeMetadata-class}
\alias{BgeeMetadata}
\title{BgeeMetadata S4 class}
\description{
An S4 class that contains all information to retrieve
intergenic regions generated by Bgee.
}
\section{Slots}{

\describe{
\item{\code{intergenic_release}}{Bgee intergenic release that will be used}

\item{\code{all_releases}}{list of all reference intergenic releases that can be
used to generate your present/absent expression calls.}

\item{\code{intergenic_prefix}}{String used to generate an intergenic release
specific output directory}

\item{\code{custom_intergenic_path}}{path to a local version of reference intergenic
fasta file. If NULL (by default) the reference intergenic fasta file will be
downloaded. If not NULL BgeeCall will merge this local reference intergenic
file with the transcriptome. Except if you generated your own intergenic
regions always keep it NULL.}
}}

