% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kallisto.R
\name{create_kallisto_index}
\alias{create_kallisto_index}
\title{Create kallisto indexes.}
\usage{
create_kallisto_index(myKallistoMetadata, myBgeeMetadata, myUserMetadata)
}
\arguments{
\item{myKallistoMetadata}{A Reference Class KallistoMetadata object.}

\item{myBgeeMetadata}{A Reference Class BgeeMetadata object.}

\item{myUserMetadata}{A Reference Class UserMetadata object.}
}
\description{
This function creates kallisto indexes. 
Two indexes are created. One with default kmer value and one with kmer size of 21nt.
}
\author{
Julien Wollbrett.
}
