% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_fasta_intergenic}
\alias{download_fasta_intergenic}
\title{Download fasta intergenic}
\usage{
download_fasta_intergenic(myBgeeMetadata = new("BgeeMetadata"),
  myUserMetadata, bgee_intergenic_file)
}
\arguments{
\item{myBgeeMetadata}{A Reference Class BgeeMetadata object (optional)}

\item{myUserMetadata}{A Reference Class UserMetadata object.}

\item{bgee_intergenic_file}{path where intergenic file will be saved}
}
\description{
Check if Bgee reference intergenic fasta file has already 
been downloaded. If not the file is downloaded.
}
\examples{
{
bgee_intergenic_file <- file.path(getwd(), "intergenic.fasta")
userMetadata <- new("UserMetadata", species_id = "7227")
}

}
