% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kallisto.R
\name{download_kallisto}
\alias{download_kallisto}
\title{Download binary version of kallisto.}
\usage{
download_kallisto(myKallistoMetadata, myUserMetadata)
}
\arguments{
\item{myKallistoMetadata}{A Reference Class KallistoMetadata object.}
}
\description{
Check your OS and download correct binary version of kallisto.
}
\author{
Julien Wollbrett.
}
