% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DropletMetadata.R
\docType{class}
\name{DropletMetadata-class}
\alias{DropletMetadata-class}
\alias{DropletMetadata}
\title{DropletMetadata s4 class}
\description{
An S4 class that contains all metadata needed to run the present/absent calls
for pseudobulked droplet-based single-cell RNA-seq data.
}
\section{Slots}{

\describe{
\item{\code{sequencing_technology}}{Character string indicating the single-cell target technology 
(e.g., "10xV2", "10xV3", "DropSeq"). Essential for bustools processing (check kallisto kb --list for 
supported technologies).}

\item{\code{celltype_annotation}}{A data.frame mapping unique cell barcodes to their 
corresponding biological groupings, such as cell types or clusters.}

\item{\code{count_matrix}}{An optional highly sparse matrix (dgCMatrix) provided by the user 
if the computationally expensive fastq processing phase is bypassed.}

\item{\code{fastq_r1_path}}{Character vector of paths to Read 1 FASTQ files (barcodes/UMIs).}

\item{\code{fastq_r2_path}}{Character vector of paths to Read 2 FASTQ files (biological sequences).}

\item{\code{h5ad_path}}{Optional character path to an AnnData (.h5ad) file containing the count matrix and metadata.}
}}

