% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galaxyWrapper.R
\name{galaxy_calls}
\alias{galaxy_calls}
\title{generate calls for galaxy}
\usage{
galaxy_calls(
  species_id = NULL,
  run_ids = NULL,
  rnaseq_lib_path = NULL,
  reads_size = 51,
  transcriptome_path = NULL,
  annotation_path = NULL,
  output_file = NULL,
  custom_intergenic_path = NULL
)
}
\arguments{
\item{species_id}{The NCBI taxon ID of the species}

\item{run_ids}{The subset of runs from the library used to generate the calls.
By default all runs are used.}

\item{rnaseq_lib_path}{Path to the directory containing the fastq files of the library}

\item{reads_size}{mean size of the reads. (Default = 51)}

\item{transcriptome_path}{Path to the transcriptome fastq file}

\item{annotation_path}{path to the gtf file containing gene annotation}

\item{output_file}{file containing the path to all result files generated}

\item{custom_intergenic_path}{(optional) To use if the "custom" reference intergenic
release has been selected. Provides the path to the reference intergenic file}
}
\description{
Function generating calls for galaxy. It is a simple adaptor of the
function ```generate_calls_workflow``` adapted to galaxy requirements.
}
