% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presenceAbsence.R
\name{generate_presence_absence}
\alias{generate_presence_absence}
\title{Generate presence absence}
\usage{
generate_presence_absence(myAbundanceMetadata, myBgeeMetadata,
  myUserMetadata)
}
\arguments{
\item{myAbundanceMetadata}{A descendant object of the Class myAbundanceMetadata.}

\item{myBgeeMetadata}{A Class BgeeMetadata object.}

\item{myUserMetadata}{A Class UserMetadata object.}
}
\description{
Generate presence absence calls
}
\author{
Julien Wollbrett

Julien Roux
}
