% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserMetadata.R
\name{getWorkingPath}
\alias{getWorkingPath}
\title{`working_path` Getter}
\usage{
getWorkingPath(userObject)
}
\arguments{
\item{userObject}{The UserMetadata object}
}
\value{
the value of the `working_path` slot of the object
}
\description{
Get value of the `working_path` slot
}
\examples{
{
user <- new("UserMetadata")
working_path <- getWorkingPath(user)
}

}
