% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareTranscriptome.R
\name{merge_transcriptome_and_intergenic}
\alias{merge_transcriptome_and_intergenic}
\title{Merge transcriptome file provided by the user with the Bgee intergenic fasta file.}
\usage{
merge_transcriptome_and_intergenic(myKallistoMetadata, myBgeeMetadata,
  myUserMetadata)
}
\arguments{
\item{myKallistoMetadata}{A Reference Class KallistoMetadata object.}

\item{myBgeeMetadata}{A Reference Class BgeeMetadata object.}

\item{myUserMetadata}{A Reference Class UserMetadata object.}
}
\description{
This function will create a file corresponding to the concatenation of the transcriptome fasta file provided by the user and the corresponding intergenic fasta file created by Bgee.
}
\author{
Julien Wollbrett.
}
