% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCallsPipeline.R
\name{run_from_dataframe}
\alias{run_from_dataframe}
\title{generate present/absent calls from a data frame}
\usage{
run_from_dataframe(myAbundanceMetadata = new("KallistoMetadata"),
  myBgeeMetadata = new("BgeeMetadata"), userMetadataDataFrame)
}
\arguments{
\item{myAbundanceMetadata}{A Reference Class BgeeMetadata object (optional)}

\item{myBgeeMetadata}{A Reference Class BgeeMetadata object (optional)}

\item{myUserMetadata}{A data frame containing all information needed to generate UserMetadata objects}
}
\description{
Main function allowing to generate present/absent calls for one library described in a data frame
}
\author{
Julien Wollbrett
}
