% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCallsPipeline.R
\name{run_from_file}
\alias{run_from_file}
\title{generate present/absent calls from a file}
\usage{
run_from_file(myAbundanceMetadata = new("KallistoMetadata"),
  myBgeeMetadata = new("BgeeMetadata"), userMetadataFile)
}
\arguments{
\item{myAbundanceMetadata}{A Reference Class BgeeMetadata object (optional)}

\item{myBgeeMetadata}{A Reference Class BgeeMetadata object (optional)}

\item{userMetadataFile}{A tsv file describing all user libraries for which 
present/absent calls have to be generated}
}
\description{
Main function allowing to generate present/absent calls for some 
libraries described in a file. Each line of the file describes one RNA-Seq library.
}
\author{
Julien Wollbrett
}
