% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCallsPipeline.R
\name{run_from_object}
\alias{run_from_object}
\title{generate present/absent calls from a UserMetadata object}
\usage{
run_from_object(myAbundanceMetadata = new("KallistoMetadata"),
  myBgeeMetadata = new("BgeeMetadata"), myUserMetadata)
}
\arguments{
\item{myAbundanceMetadata}{A Reference Class BgeeMetadata object (optional)}

\item{myBgeeMetadata}{A Reference Class BgeeMetadata object (optional)}

\item{myUserMetadata}{A Reference Class UserMetadata object.}
}
\description{
Main function allowing to generate present/absent calls for one library described as a UserMetadata object
}
\author{
Julien Wollbrett
}
