% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kallisto.R
\name{run_kallisto}
\alias{run_kallisto}
\title{Run one kallisto abundance analyse}
\usage{
run_kallisto(myKallistoMetadata, myBgeeMetadata, myUserMetadata)
}
\arguments{
\item{myKallistoMetadata}{A Reference Class KallistoMetadata object.}

\item{myBgeeMetadata}{A Reference Class BgeeMetadata object.}

\item{myUserMetadata}{A Reference Class UserMetadata object. 
This object has to be edited before running kallisto @seealso UserMetadata.R}
}
\description{
Run kallisto and all preliminary steps if needed like :
- creation of transcriptome with intergenic (if needed)
- installation of kallisto (if needed)
- index creation (if needed)
- run kallisto quantification
}
\author{
Julien Wollbrett.
}
