% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcriptToGene.R
\name{run_tximport}
\alias{run_tximport}
\title{Run tximport}
\usage{
run_tximport(myAbundanceMetadata, myBgeeMetadata, myUserMetadata)
}
\arguments{
\item{myAbundanceMetadata}{A descendant object of the Class myAbundanceMetadata.}

\item{myBgeeMetadata}{A Reference Class BgeeMetadata object.}

\item{myUserMetadata}{A Reference Class UserMetadata object.}
}
\description{
Run tximport. Will summarize abundance estimation from transcript 
level to gene level if `myAbundanceMetadata@txout == FALSE`. 
Otherwise keep abundance estimation at transcript level.
}
\author{
Julien Wollbrett
}
