% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserMetadata.R
\name{setAnnotationFromObject}
\alias{setAnnotationFromObject}
\title{Set annotation_object of one UserMetadata object}
\usage{
setAnnotationFromObject(userObject, annotationObject, annotationName)
}
\arguments{
\item{userObject}{UserMetadata object}

\item{annotationObject}{Object of the GRanges S4 class}

\item{annotationName}{Name of the annotation. Will be used to create annotation folders.}
}
\value{
an object of UserMetadata
}
\description{
Method of the class UserMetadata. Set annotation_object of one UserMetadata object 
by using one GRanges object as input.
}
\details{
GRanges object is used as input. The class GRanges comes from the rtacklayer package
}
