% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserMetadata.R
\docType{methods}
\name{setRNASeqLibPath}
\alias{setRNASeqLibPath}
\alias{setRNASeqLibPath,UserMetadata,character-method}
\alias{setRNASeqLibPath,userMetadata,character}
\title{`rnaseq_lib_path` Setter}
\usage{
setRNASeqLibPath(userObject, rnaSeqLibPath)

\S4method{setRNASeqLibPath}{UserMetadata,character}(userObject,
  rnaSeqLibPath)
}
\arguments{
\item{userObject}{The UserMetadata object}

\item{rnaSeqLibPath}{path to the directory wanted as `rnaseq_lib_path`}
}
\value{
An object of the class UserMetadata with new `rnaseq_lib_path`
 value
}
\description{
Set value of the `rnaseq_lib_path` slot
}
\examples{
{
user <- new("UserMetadata")
user <- setRNASeqLibPath(user, getwd())
}

}
