% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserMetadata.R
\name{setTranscriptomeFromFile}
\alias{setTranscriptomeFromFile}
\title{Set transcriptome_object of one UserMetadata object}
\usage{
setTranscriptomeFromFile(userObject, transcriptomePath, transcriptomeName)
}
\arguments{
\item{userObject}{The UserMetadata object}

\item{annotationPath}{Absolute path to the transcriptome file}

\item{annotationName}{(optional) Name of the trancriptome Will be used to create folders.}
}
\value{
An object of the class UserMetadata
}
\description{
Method of the class UserMetadata. Set transcriptome_object of one UserMetadata object 
by providing the path to a fasta transcriptome file.
}
\details{
If no annotationNAme is provided the name of the annotation file will be used to create folders.
}
