% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserMetadata.R
\docType{methods}
\name{setTranscriptomeFromFile}
\alias{setTranscriptomeFromFile}
\alias{setTranscriptomeFromFile,UserMetadata,character,character-method}
\alias{setTranscriptomeFromFile,userMetadata,character,character}
\title{Set transcriptome_object of one UserMetadata object}
\usage{
setTranscriptomeFromFile(userObject, transcriptomePath, transcriptomeName)


  \S4method{setTranscriptomeFromFile}{UserMetadata,character,character}(userObject,
  transcriptomePath, transcriptomeName = "")
}
\arguments{
\item{userObject}{The UserMetadata object}

\item{transcriptomePath}{Absolute path to the transcriptome file}

\item{transcriptomeName}{(optional) Name of the trancriptome. Will be used to 
create folders.}
}
\value{
An object of the class UserMetadata
}
\description{
Method of the class UserMetadata. Set transcriptome_object of 
one UserMetadata object 
by providing the path to a fasta transcriptome file.
}
\details{
If no transcriptomeName is provided the name of the transcriptome file 
will be used to create folders.
}
\examples{
{
library(AnnotationHub)
ah <- AnnotationHub()
ah_resources <- AnnotationHub::query(ah, c("Ensembl", "Caenorhabditis elegans", "84"))
annotation_object <- ah_resources[["AH50789"]]
transcriptome_object <- rtracklayer::import.2bit(ah_resources[["AH50453"]])
transcriptome_path <- file.path(getwd(), "transcriptome.fa")
writeXStringSet(transcriptome_object, transcriptome_path)
user <- new("UserMetadata")
user <- setTranscriptomeFromFile(user, transcriptome_path,
                                 "transcriptome_name")
}

}
